+++ apps/coldsurf-io/features/index.ts
@@ -1,2 +1,3 @@
 export * from './link';
 export * from './notion';
+export * from './share';